requirements.txt
torch>=1.10
transformers>=4.15
Pillow
tk
requests
